+++ src/plone/app/mosaic/browser/static/js/mosaic.layout.js
@@ -129,6 +129,11 @@ define([
                 elm = e.srcElement;
             }
 
+            // If clicked TinyMCE toolbar
+            if ($(elm).parents(".mce-panel").length > 0) {
+                return;
+            }
+
             // If clicked outside a tile
             if ($(elm).parents(".mosaic-tile").length === 0) {
 
@@ -1336,8 +1341,7 @@ define([
 
         // Re-init rich text editor after tile has been moved in DOM
         $(".mosaic-new-tile .mosaic-rich-text").each(function () {
-            if (tinyMCE && tinyMCE.get($(this).attr("id"))) {
-                tinyMCE.get($(this).attr("id")).remove();
+            if ($(this).data("pattern-tinymce")) {
                 $(this).mosaicEditor();
             }
         });